backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Marshmallow==0.15.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
python-dotenv==1.0.0
PyMySQL==1.1.0
cryptography==41.0.4
bcrypt==4.0.1
requests==2.31.0
gunicorn==21.2.0
